Move IRemoteHostUtils to common package
