+++ game-core/src/main/java/games/strategy/engine/lobby/common/IRemoteHostUtils.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.lobby.server;
+package games.strategy.engine.lobby.common;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -6,6 +6,10 @@ import games.strategy.engine.message.IRemote;
 import games.strategy.engine.message.RemoteName;
 import games.strategy.net.INode;
 
+/**
+ * A service that provides operations to manage a headless game server (bot) connected to the lobby. Each headless game
+ * server provides this service. Only the lobby server is permitted to invoke the operations of this service.
+ */
 public interface IRemoteHostUtils extends IRemote {
   String getConnections();
