+++ game-core/src/main/java/games/strategy/engine/lobby/server/RemoteHostUtils.java
@@ -3,6 +3,7 @@ package games.strategy.engine.lobby.server;
 import java.util.Set;
 
 import games.strategy.engine.framework.headlessGameServer.HeadlessGameServer;
+import games.strategy.engine.lobby.common.IRemoteHostUtils;
 import games.strategy.engine.message.MessageContext;
 import games.strategy.net.INode;
 import games.strategy.net.IServerMessenger;
