+++ lobby/src/main/java/games/strategy/engine/lobby/server/ModeratorController.java
@@ -7,6 +7,7 @@ import javax.annotation.Nullable;
 
 import games.strategy.engine.config.lobby.LobbyPropertyReader;
 import games.strategy.engine.lobby.common.IModeratorController;
+import games.strategy.engine.lobby.common.IRemoteHostUtils;
 import games.strategy.engine.lobby.common.LobbyConstants;
 import games.strategy.engine.lobby.server.db.BannedMacController;
 import games.strategy.engine.lobby.server.db.BannedUsernameController;
